### 2024-07-19 #7011 @editorialbot
[PRE REVIEW]: CompressedBeliefMDPs.jl: A Julia Package for Solving Large POMDPs with Belief Compression
**Submitting author:** @FlyingWorkshop (<a href="http://orcid.org/0009-0001-3978-9462">Logan Bhamidipaty</a>)
**Repository:** https://github.com/JuliaPOMDP/CompressedBeliefMDPs.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
