### 2024-07-20 #7011 @crvernon
**Note**:  This was previously submitted and rejected here:  https://github.com/openjournals/joss-reviews/issues/6673

The authors have made modifications to the codebase that now qualify this submission for consideration.
