### 2024-08-03 #7011 @FlyingWorkshop
Hi @bmcfee! Thanks for editing this paper. Here's a list of 5 reviewers I think would be well-suited to the paper:

1. [Sandesh Katakam](https://reviewers.joss.theoj.org/reviewers/2792): Has reinforcement learning (RL) experience and knows Julia.
2. [Lokesh Mohanty](https://reviewers.joss.theoj.org/reviewers/4361): Has RL experience. Doesn't list Julia as a language, but the syntax is similar enough to Python that I don't think it would be an issue.
3. [Mariana Meireles](https://reviewers.joss.theoj.org/reviewers/4160): Similarly has RL and Python experience, but doesn't list Julia.
4. [Lili Wu](https://reviewers.joss.theoj.org/reviewers/3143): Experience with decision-making, RL, and Python, but doesn't list Julia.
5. [Manuel Razo-Mejia](https://reviewers.joss.theoj.org/reviewers/4076): Lists Julia and has experience with machine learning (ML).

Happy to assist further if you'd like.
