+++ scripts/README.md
@@ -1,0 +1,1 @@
+Requires installing `Plots.jl`
