+++ Project.toml
@@ -1,7 +1,7 @@
 name = "CompressedBeliefMDPs"
 uuid = "0a809e47-b8eb-4578-b4e8-4c2c5f9f833c"
 authors = ["Logan-Mondal-Bhamidipaty"]
-version = "1.1.1"
+version = "1.1.2"
 
 [deps]
 Bijections = "e2ed5e7c-b2de-5872-ae92-c73ca462fb04"
@@ -18,10 +18,7 @@ NearestNeighbors = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
 POMDPModels = "355abbd5-f08e-5560-ac9e-8b5f2592a0ca"
 POMDPTools = "7588e00f-9cae-40de-98dc-e0c70c48cdd7"
 POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 ParticleFilters = "c8b314e2-9260-5cf8-ae76-3be7461ca6d0"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
@@ -38,11 +35,8 @@ MultivariateStats = "0.9, 0.10"
 NearestNeighbors = "0.4"
 POMDPModels = "0.4.20"
 POMDPTools = "0.1.6, 1"
-POMDPs = "0.9.6, 1"
-Parameters = "0.12"
+POMDPs = "1"
 ParticleFilters = "0.5.6"
-Plots = "1"
-ProgressMeter = "1"
 StaticArrays = "1.9.3"
 julia = "1"
 
