+++ v1.1.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/#API-Documentation","page":"API Documentation","title":"API Documentation","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"CurrentModule = CompressedBeliefMDPs","category":"page"},{"location":"api/#Contents","page":"API Documentation","title":"Contents","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Pages = [\"api.md\"]","category":"page"},{"location":"api/#Index","page":"API Documentation","title":"Index","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Pages = [\"api.md\"]","category":"page"},{"location":"api/#Types/Functors","page":"API Documentation","title":"Types/Functors","text":"","category":"section"},{"location":"api/","page":"API Documentation","title":"API Documentation","text":"Sampler\nCompressor\nCompressedBeliefMDP\nCompressedBeliefPolicy\nCompressedBeliefSolver","category":"page"},{"location":"api/#CompressedBeliefMDPs.Sampler","page":"API Documentation","title":"CompressedBeliefMDPs.Sampler","text":"Abstract type for an object that defines how the belief should be sampled.\n\n\n\n\n\n","category":"type"},{"location":"api/#CompressedBeliefMDPs.Compressor","page":"API Documentation","title":"CompressedBeliefMDPs.Compressor","text":"Abstract type for an object that defines how the belief should be compressed.\n\n\n\n\n\n","category":"type"},{"location":"api/#CompressedBeliefMDPs.CompressedBeliefMDP","page":"API Documentation","title":"CompressedBeliefMDPs.CompressedBeliefMDP","text":"CompressedBeliefMDP{B, A}\n\nThe CompressedBeliefMDP struct is a generalization of the compressed belief-state MDP presented in  Exponential Family PCA for Belief Compression in POMDPs.\n\nType Parameters\n\nB: The type of compressed belief states.\nA: The type of actions.\n\nFields\n\nbmdp::GenerativeBeliefMDP: The generative belief-state MDP.\ncompressor::Compressor: The compressor used to compress belief states.\nϕ::Bijection: A bijection representing the mapping from uncompressed belief states to compressed belief states. See notes. \n\nConstructors\n\nCompressedBeliefMDP(pomdp::POMDP, updater::Updater, compressor::Compressor)\nCompressedBeliefMDP(pomdp::POMDP, sampler::Sampler, updater::Updater, compressor::Compressor)\n\nConstructs a CompressedBeliefMDP using the specified POMDP, updater, and compressor.\n\nwarning: Warning\nThe 4-argument constructor is a quality-of-life constructor that calls fit! on the given compressor. \n\nExample Usage\n\npomdp = TigerPOMDP()\nupdater = DiscreteUpdater(pomdp)\ncompressor = PCACompressor(1)\nmdp = CompressedBeliefMDP(pomdp, updater, compressor)\n\nFor continuous POMDPs, see ParticleFilters.jl.\n\nNotes\n\nWhile compressions aren't usually injective, we cache beliefs and their compressions on a first-come, first-served basis, so we can effectively use a bijection without loss of generality.\n\n\n\n\n\n","category":"type"},{"location":"api/#CompressedBeliefMDPs.CompressedBeliefPolicy","page":"API Documentation","title":"CompressedBeliefMDPs.CompressedBeliefPolicy","text":"CompressedBeliefPolicy\n\nMaps a base policy for the compressed belief-state MDP to a policy for the true POMDP.\n\nFields\n\nm::CompressedBeliefMDP: The compressed belief-state MDP.\nbase_policy::Policy: The base policy used for decision-making in
[…]
