+++ versions.js
@@ -4,5 +4,5 @@ var DOC_VERSIONS = [
   "v1.0",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v1.1.0";
+var DOCUMENTER_NEWEST = "v1.1.2";
 var DOCUMENTER_STABLE = "stable";
