### 2024-08-19 #7011 @FlyingWorkshop
I added a minor patch to update some old dependencies, please refer to [v1.1.2](https://github.com/JuliaPOMDP/CompressedBeliefMDPs.jl/releases/tag/v1.1.2) instead of v1.1.1 @AlCap23, @bmcfee.
