### 2024-08-25 #7011 @FlyingWorkshop
Thank you @AlCap23 and @bmcfee for agreeing to review our submission. I’m reaching out to PhD programs this coming month, so if there’s anything I can do to help expedite this review please let me know and I’d be happy to address them
