### 2024-09-23 #7011 @FlyingWorkshop
Hi @AlCap23 and @bmcfee, sending another ping! I'm writing another paper and would like to cite this project. Would it be possible for you to move the paper to "Review" from "Pre-Review," so that I can cite the provisional DOI? Please let me know if I can help expedite the process in any way.
