### 2024-10-11 #7011 @bmcfee
Hey, just checking back in here - still working on reviewers in the background here.

To your question, I'm not sure if starting the review would provision the DOI immediately, but sure we can start the review process and add the second reviewer upon confirmation.  Thanks for your patience!
