### 2024-10-11 #7346 @editorialbot
[REVIEW]: CompressedBeliefMDPs.jl: A Julia Package for Solving Large POMDPs with Belief Compression
**Submitting author:** @FlyingWorkshop (<a href="http://orcid.org/0009-0001-3978-9462">Logan Bhamidipaty</a>)
**Repository:** https://github.com/JuliaPOMDP/CompressedBeliefMDPs.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.14
