### 2024-11-12 #7346 @AlCap23
Hey! I'm sorry for being absent so long. I had something to take care of privately and felt under the weather.

This is an excellent paper and software contribution to Julia's POMDP ecosystem.

Regarding the overall scope and target of this contribution, I have little to add, but there are some things:

Minor issues for the publication:

1. The missing DOI of `Perseus: Randomized point-based value iteration for POMDPs` can be completed using the current `joi` field in the `paper.bib`.
2. I like the example overview, which showcases the flexibility and solver-agnostic behavior. However, to make it copy-pasteable, I suggest rewriting one of the examples to use a specific solver. Given that you have many in the Readme, this would be an excellent addition 👍 .

Repository / Source Code:
1. Minor Issue: The `Continuous Example` in the docs is doubled
2. I don't know if the POMDP ecosystem has explicit contribution guidelines, but linking them explicitly in the documentation would be helpful if this is the case. Once this is double-checked, I'll check the corresponding item in the review list.

I also have a recommendation based on my experience: Use `StableRNGs` in the tests. While not entirely foolproof, this method saved me a lot of time dealing with consistent testing behavior 😄 .

I still have to check the reference syntax for completeness.
