+++ paper.bib
@@ -151,7 +151,7 @@ series = {IJCAI'03}
   volume={24},
   pages={195--220},
   year={2005},
-  joi={10.1613/jair.1659}
+  doi={10.1613/jair.1659}
 }
 
 @article{hsvi,
