### 2024-11-17 #7346 @FlyingWorkshop
Thank you again for the review @AlCap23! I've just pushed updates based on your feedback:

1. Removed the redundant `Continuous Example` from the docs
2. Linked the **contribution guidelines** for POMDPs.jl in the [documentation](https://juliapomdp.github.io/CompressedBeliefMDPs.jl/dev/#Contribution-Guidelines).
3. Changed `joi` to `doi` in `paper.bib`
4. We opted not to include a specific solver in the snippet for the following reasons: (a) to avoid introducing unrelated solvers into the paper, (b) to highlight that any solver can be used, and (c) because the signature for constructing an MDP solver can be complex and might overly lengthen the example. However, if you feel strongly about this, I'm happy to discuss this point further.

Let me know if you have any questions about the reference syntax 😄 .
