### 2024-11-18 #7346 @AlCap23
@bmcfee I've finished the review, and I have nothing to add. The tests are running (I double-checked locally using Julia Version 1.10.4 ), and the Documentation is up and running andes, and includes both examples and an API description. The contribution guide references the "parent" ecosystem, and references for more advanced (sub) solvers are given ( Crux.jl as an example for incorporating neural network-based solvers ).
