### 2024-11-19 #7346 @FlyingWorkshop
@bmcfee, I noticed on the [JOSS website that there has to be two reviewers](https://joss.readthedocs.io/en/latest/editing.html#:~:text=JOSS%20papers%20have%20to%20have%20a%20minimum%20of%20two%20reviewers%20per%20submission) for a paper to be accepted. Has there been any progress on finding a second reviewer?

As a reminder, I [listed some people in the pre-review thread](https://github.com/openjournals/joss-reviews/issues/7011#issuecomment-2267166732) who I think would be good fits to review this project. There is also another list of reviewers I created for a related project that I also submitted to JOSS [here](https://github.com/openjournals/joss-reviews/issues/7353#issuecomment-2429385553). I believe they too are good candidates to review CompressedBeliefMDPS.jl.
