### 2025-04-29 #7346 @mrazomej
Hi @crvernon ,

Unfortunately, this is not a good time for me. The package looks really interesting, but I cannot commit at this time to put in the necessary time to review it carefully. I sincerely apologize.
