### 2025-04-29 #7346 @lwu9
Hi there,

Thanks for the invite, but I don't know Julia so I don't think I'm a good
fit.

Best,
Lili
