### 2025-05-05 #7346 @marimeireles
Hey folks, while trying to install the package I got a segfault.

As you can see from my paths I'm using mamba. Is there any known issue in the conda <> julia ecosystem?

```bash
Precompiling project...
┌ Error: curl_multi_assign: 1                        ]  55/218
└ @ Downloads.Curl ~/.local/share/mamba/envs/julia/share/julia/stdlib/v1.10/Downloads/src/Curl/utils.jl:57
┌ Error: curl_multi_assign: 1
└ @ Downloads.Curl ~/.local/share/mamba/envs/julia/share/julia/stdlib/v1.10/Downloads/src/Curl/utils.jl:57
┌ Error: curl_multi_assign: 1
└ @ Downloads.Curl ~/.local/share/mamba/envs/julia/share/julia/stdlib/v1.10/Downloads/src/Curl/utils.jl:57
  Progress [===========>                             ]  55/218
  ✓ NaNMath
  ◒ MLStyle
  ✓ Crayons
  ✓ BitFlags
  ◑ LoggingExtras
  ◐ Functors
  ◓ StaticArraysCore
  ◐ Bijections
  ◐ Baselet
  ◓ SuiteSparse
  ◓ Statistics
  ◒ MacroTools
  ◑ Preferences
  ◒ InverseFunctions → InverseFunctionsTestExt
  ◐ NameResolution
  ◓ Compat → CompatLinearAlgebraExt

[45297] signal (11.1): Segmentation fault: 11
in expression starting at none:0
curl_multi_socket_action at /Users/mariana/.local/share/mamba/envs/julia/lib/libcurl.4.dylib (unknown line)
curl_multi_socket_action at /Users/runner/miniforge3/conda-bld/julia_1718579755449/work/usr/share/julia/stdlib/v1.10/Downloads/src/Curl/Curl.jl:51 [inlined]
curl_multi_socket_action at /Users/runner/miniforge3/conda-bld/julia_1718579755449/work/usr/share/julia/stdlib/v1.10/Downloads/src/Curl/Curl.jl:59 [inlined]
macro expansion at /Users/runner/miniforge3/conda-bld/julia_1718579755449/work/usr/share/julia/stdlib/v1.10/Downloads/src/Curl/utils.jl:56 [inlined]
#59 at /Users/runner/miniforge3/conda-bld/julia_1718579755449/work/usr/share/julia/stdlib/v1.10/Downloads/src/Curl/Multi.jl:199
lock at ./lock.jl:229
#58 at /Users/runner/miniforge3/conda-bld/julia_1718579755449/work/usr/share/julia/stdlib/v1.10/Downloads/src/Curl/Multi.jl:197
jfptr_YY.58_92848.1 at /Users/mariana/.local/share/mamba/envs/julia/lib/julia/sys.dylib (unknown line)
_jl_invoke at /usr/local/src/conda/julia-1.10.4/src/gf.c:0 [inlined]
ijl_apply_generic at /usr/local/src/conda/julia-1.10.4/src/gf.c:3077
jl_apply at /usr/local/src/conda/julia-1.10.4/src/./julia.h:1982 [inlined]
start_task at /usr/local/src/conda/julia-1.10.4/src/task.c:1238
Allocations: 16713978 (Pool: 16682663; Big: 31315); GC: 26
[1]    45297 segmentation fault  julia
```

Will try later with `brew`.

best,
