### 2025-05-06 #7346 @FlyingWorkshop
Hi @marimeireles, thanks so much for agreeing to review our project!

Have you already installed Julia on your machine? I just tried re-installing the package didn't encounter any bugs. I'm not very familiar with mamba, but if you're running into issues, you could try `brew` (as you suggest) or the `curl` command on this [installation guide](https://docs.julialang.org/en/v1/manual/installation/). Let me know if you encounter any more issues.
