+++ paper.bib
@@ -15,6 +15,8 @@ numpages = {5},
 keywords = {sequential decision making, open-source, POMDP, MDP, Julia}
 }
 
+@article{Bhamidipaty2025, doi = {10.21105/joss.07403}, url = {https://doi.org/10.21105/joss.07403}, year = {2025}, publisher = {The Open Journal}, volume = {10}, number = {105}, pages = {7403}, author = {Logan Mondal Bhamidipaty and Mykel J. Kochenderfer and Trevor Hastie}, title = {ExpFamilyPCA.jl: A Julia Package for Exponential Family Principal Component Analysis}, journal = {Journal of Open Source Software} }
+
 @article{Roy,
    title={Finding Approximate POMDP solutions Through Belief Compression},
    volume={23},
