### 2025-06-11 6ef64b8
bib tex update
