### 2025-06-16 #7346 @FlyingWorkshop
Working through the checklist now.

> * [ ]  LINE 120:  The "j" in "julia" should be capitalized.  You can do this in your bib file by placing the characters you wish to maintain the formatting for in curly brackets.
>   [ ]   Confirm that the title of your paper in your paper.md file has no special formatting (e.g., backticks) as this will cause formatting issues when accepted.
>   [ ]   Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on code repository.
>   [ ]   Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).
>   [ ]   Please respond with the DOI of the archived version here along with the version number released.
