### 2025-06-16 fbac2f1
minor documentation updates
