### 2025-06-16 590485b
remove vestigial dependency
