+++ src/CompressedBeliefMDPs.jl
@@ -13,7 +13,7 @@ using Distributions
 using Bijections
 using NearestNeighbors
 using StaticArrays
-using Plots
+# using Plots
 # using ProgressMeter
 
 using LinearAlgebra
