+++ Project.toml
@@ -19,7 +19,6 @@ POMDPModels = "355abbd5-f08e-5560-ac9e-8b5f2592a0ca"
 POMDPTools = "7588e00f-9cae-40de-98dc-e0c70c48cdd7"
 POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
 ParticleFilters = "c8b314e2-9260-5cf8-ae76-3be7461ca6d0"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
@@ -38,7 +37,6 @@ POMDPModels = "0.4.20"
 POMDPTools = "0.1.6, 1"
 POMDPs = "1"
 ParticleFilters = "0.5.6"
-Plots = "1.40.14"
 StaticArrays = "1.9.3"
 julia = "1"
 
