+++ Project.toml
@@ -1,7 +1,7 @@
 name = "CompressedBeliefMDPs"
 uuid = "0a809e47-b8eb-4578-b4e8-4c2c5f9f833c"
 authors = ["Logan-Mondal-Bhamidipaty"]
-version = "1.1.2"
+version = "1.1.3"
 
 [deps]
 Bijections = "e2ed5e7c-b2de-5872-ae92-c73ca462fb04"
