### 2025-06-20 #7346 @crvernon
🥳 Congratulations on your new publication @FlyingWorkshop! Many thanks to @AlCap23 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
